mongo fix

+++ backend/src/test/java/com/example/safestorage/MongoTests.java
@@ -3,6 +3,7 @@ package com.example.safestorage;
 import com.example.safestorage.models.Note;
 import com.example.safestorage.models.User;
 import com.example.safestorage.services.*;
+import org.junit.After;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -19,6 +20,8 @@ class MongoTests {
 
     private final MongoTemplate template;
 
+    private User entity;
+
     @Autowired
     MongoTests(NoteService noteService, UserService userService, MongoTemplate template) {
         this.noteService = noteService;
@@ -33,12 +36,9 @@ class MongoTests {
         var pass = "ewf";
         var user = new User(name, pass);
         userService.saveUser( user );
-        var entity = userService.findUserByName( name );
+        entity = userService.findUserByName( name );
         assert name.equals( entity.getUsername() );
         assert pass.equals( entity.getPasswordHash() );
-
-        //TODO: move to after test
-        userService.removeUser( entity.getId() );
     }
 
     @Test
@@ -48,12 +48,30 @@ class MongoTests {
         var user = new User(name, pass);
         userService.saveUser( user );
         var query = new Query();
-        var entity = template.findOne( query, User.class );// userService.findUserByName( name );
+        entity = template.findOne( query, User.class );// userService.findUserByName( name );
         query.fields().exclude( "username" );
         assert entity.getUsername().equals( name );
         var entity2 = template.findOne( query, User.class );// userService.findUserByName( name );
         assert entity2.getUsername() == null;
-        //TODO: move to after test
-        userService.removeUser( entity.getId() );
+    }
+
+    @Test
+    void getUserTest() {
+        var entity = template.findById( "5f1ec5cdefd8ff5e799e6c94", User.class );
+        System.out.println(entity);
+        assert entity != null;
+    }
+
+    @Test
+    void getNoteTest() {
+        var entity = template.findById( "5f1ec8acf87dff6bc91bad21", Note.class );
+        System.out.println(entity);
+        assert entity != null;
+    }
+
+    @After
+    public void cleaning() {
+        if(entity != null)
+            userService.removeUser( entity.getId() );
     }
 }
