+++ backend/src/test/java/com/example/safestorage/NoteTests.java
@@ -2,8 +2,8 @@ package com.example.safestorage;
 
 import com.example.safestorage.models.Note;
 import com.example.safestorage.services.NoteService;
-import com.example.safestorage.services.NoteServiceImpl;
 import com.example.safestorage.services.UserService;
+import org.junit.After;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -20,6 +20,8 @@ public class NoteTests {
 
     private final MongoTemplate template;
 
+    private Note entity;
+
     @Autowired
     NoteTests(NoteService noteService, UserService userService, MongoTemplate template) {
         this.noteService = noteService;
@@ -31,9 +33,15 @@ public class NoteTests {
     void getDetailsTest() {
         var note = new Note(new byte[] {4}, new byte[] {4, 5}, "123", null);
         noteService.saveNote( note );
-        var entity = template.findOne( Query.query( Criteria.where( "ownerId" ).is(note.getOwnerId()) ), Note.class );
+        entity = template.findOne( Query.query( Criteria.where( "ownerId" ).is(note.getOwnerId()) ), Note.class );
+        System.out.println(entity.getId());
         var entity2 = noteService.getNoteDetails( entity.getId() );
-        assert entity.equals( entity2 );
-        noteService.removeNote( entity.getId() );
+        assert entity2.equals( entity );
+    }
+
+    @After
+    public void cleaning() {
+        if(entity != null)
+            noteService.removeNote( entity.getId() );
     }
 }
